fix: remove bad state in child component
